# swagger.yml
swagger: '2.0'
info:
  description: Webservices to upload and download files from cloud storage.
  version: 1.0.0
  title:  Cloud File Storage
host: >-
  https://magicstore.appspot.com
basePath: '/'
schemes:
  - https
paths:
  '/upload/file':
    post:
      tags:
        - uploadFile
      summary: Uploads file to google cloud storage and file details to google data store
      produces:
      - "application/json" 
      parameters:
        - name: file
          in: body
          description: file to be uploaded
          required: true
          type: string
      requestBody:
        content: 
          multipart/form-data: 
          schema:            # Request payload
            type: object
            properties:      
              file:           
                type: string
                format: binary
      responses:
        '201':
          description: File Successful uploaded
          schema:
            items:
              $ref: '#/definitions/responseUpload'
        '400':
          description: Bad request
        '404':
          description: File Not Found
        '500':
          description: Internal server Error
  '/download/file/:filename':
    post:
      tags:
        - downloadFile
      summary: Downloads file from cloud if the user has permissions
      produces:
      - "application/json" 
      parameters:
        - name: userId
          in: query
          description: User ID
          required: true
          type: string
      responses:
        '200':
          description: File Downloaded successfully
          schema:
            items:
              $ref: '#/definitions/responseDownload'
        '400':
          description: Bad request
          schema:
            items:
              $ref: '#/definitions/responseFailed'
        '404':
          description: File Not Found
          schema:
            items:
              $ref: '#/definitions/responseFailed'
        '500':
          description: Internal server Error
  '/heartbeat':
    get:
      summary: Health end point
      produces:
      - "application/json" 
      responses:
        '200':
          description: Successful operation
        '404':
          description: Not Found
        '500':
          description: Internal server Error        
  '/allFiles':
    get:
      tags:
        - allFiles
      summary: Gets list of all  files in store if user id is not passed. If user id is passed, gets list of file names uploaded by specific user.
      produces:
      - "application/json"   
      parameters:
        - name: userId
          in: query
          description: User ID
          required: false
          type: string
      responses:
        '200':
          description: Successful operation
          schema:
            items:
              $ref: '#/definitions/responseAllFiles'
        '400':
          description: Bad request
        '404':
          description: Not Found
        '500':
          description: Internal server Error

definitions:
  responseUpload:
      type: object
      properties:
        success:
          type: string
        message:
          type: string
          description: upload message
        fileIdentifier:
          type: string
          description: Unique Identifier generated on successful file upload.
  responseDownload:
      type: object
      properties:
        filename:
          type: object
          description: Contents of the file.
  responseFailed:
      type: object
      properties:
        filename:
          type: string
        errorMessage:
          type: string
          description: Description of error
  responseAllFiles:
      type: object
      properties:
        filename:
          type: string
          description: File Name
        fileUrl:
          type: string
          description: Google Cloud Storage URL where file is stored
        userId:
          type: string
          description: User ID of the user who uploaded file.
        fileId:
          type: string
          description: Unique File Identifier generated for every uploaded file.
        fileSize:
          type: string
          description: Size of the file.
        id:
          type: string
          description: ID associated with each file internally generated by datastore.